+++ src/WordFormatter/Unformatter.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Stadly\PasswordPolice\WordFormatter;
 
+use Stadly\PasswordPolice\WordFormatter;
 use Traversable;
 
-final class Unformatter extends ChainableFormatter
+final class Unformatter implements WordFormatter
 {
+    use FormatterChaining;
+
     /**
      * @param iterable<string> $words Words.
      * @return Traversable<string> The same words.